Conflicts:
	DB/DBB/standard/business/src/main/java/jp/co/ndensan/reams/db/dbb/business/core/basic/FukaBuilder.java
	DB/DBB/standard/business/src/main/java/jp/co/ndensan/reams/db/dbb/business/core/basic/HokenryoRank.java
	DB/DBB/standard/business/src/main/java/jp/co/ndensan/reams/db/dbb/business/core/basic/RankJoho.java
	DB/DBB/standard/business/src/test/java/jp/co/ndensan/reams/db/dbb/business/core/basic/FukaBuilderTest.java
	DB/DBB/standard/business/src/test/java/jp/co/ndensan/reams/db/dbb/business/core/basic/FukaTest.java
	DB/DBB/standard/business/src/test/java/jp/co/ndensan/reams/db/dbb/business/core/basic/GemmenTest.java
	DB/DBB/standard/entity/src/test/java/jp/co/ndensan/reams/db/dbb/entity/basic/helper/DbT2002FukaEntityGenerator.java
	DB/DBB/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2001ChoshuHohoDac.java
	DB/DBB/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2002FukaDac.java
	DB/DBB/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2011RankJohoDac.java
	DB/DBB/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2012HokenryoRankDac.java
	DB/DBB/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbb/persistence/db/mapper/relate/choshuyuyo/IChoshuYuyoMapper.java
	DB/DBB/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbb/persistence/db/mapper/relate/fuka/IFukaMapper.java
	DB/DBB/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2001ChoshuHohoDacTest.java
	DB/DBB/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2002FukaDacTest.java
	DB/DBB/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2003KibetsuDacTest.java
	DB/DBB/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2011RankJohoDacTest.java
	DB/DBB/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2012HokenryoRankDacTest.java
	DB/DBB/standard/service/src/main/java/jp/co/ndensan/reams/db/dbb/service/core/basic/FukaManager.java
	DB/DBB/standard/service/src/test/java/jp/co/ndensan/reams/db/dbb/service/core/basic/FukaManagerTest.java
	DB/DBC/standard/definition/pom.xml
	DB/DBC/standard/divcontroller/src/main/java/jp/co/ndensan/reams/db/dbc/divcontroller/entity/parentdiv/DBC0410012/DBC0410012TransitionEventName.java
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0410012.raml
	DB/DBX/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbx/entity/db/basic/DbT7130KaigoServiceShuruiEntity.java
	DB/DBX/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbx/entity/db/basic/DbT7131KaigoServiceNaiyouEntity.java
	DB/DBX/standard/entity/src/test/java/jp/co/ndensan/reams/db/dbx/entity/db/basic/DbT7130KaigoServiceShuruiEntityGenerator.java
	DB/DBX/standard/entity/src/test/java/jp/co/ndensan/reams/db/dbx/entity/db/basic/DbT7131KaigoServiceNaiyouEntityGenerator.java
	DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/HihokenshaDaichoBuilder.java
	DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/basic/KaigoSetai.java
	DB/DBZ/standard/business/src/test/java/jp/co/ndensan/reams/db/dbz/business/core/basic/KaigoSetaiBuilderTest.java
	DB/DBZ/standard/business/src/test/java/jp/co/ndensan/reams/db/dbz/business/core/basic/KaigoSetaiIdentifierTest.java
	DB/DBZ/standard/business/src/test/java/jp/co/ndensan/reams/db/dbz/business/core/basic/KaigoSetaiTest.java
	DB/DBZ/standard/definition/src/test/java/jp/co/ndensan/reams/db/dbz/definition/core/valueobject/hokensha/HokenshaNameTest.java
	DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT1007KyokaisoHokenryoDankaiEntity.java
	DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT2008ShotokuKanri.csv
	DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT2008ShotokuKanri.java
	DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT2008ShotokuKanriEntity.java
	DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT7014KaigoSetai.java
	DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT7014KaigoSetaiEntity.java
	DB/DBZ/standard/entity/src/test/java/jp/co/ndensan/reams/db/dbz/entity/basic/helper/DbT2008ShotokuKanriEntityGenerator.java
	DB/DBZ/standard/entity/src/test/java/jp/co/ndensan/reams/db/dbz/entity/basic/helper/DbT7014KaigoSetaiEntityGenerator.java
	DB/DBZ/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbz/persistence/db/basic/DbT1001HihokenshaDaichoDac.java
	DB/DBZ/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbz/persistence/db/basic/DbT7014KaigoSetaiDac.java
	DB/DBZ/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbz/persistence/db/basic/DbT7014KaigoSetaiDacTest.java
	DB/DBZ/standard/service/src/test/java/jp/co/ndensan/reams/db/dbz/service/core/basic/KaigoSetaiManagerTest.java

+++ DB/DBB/standard/business/src/main/java/jp/co/ndensan/reams/db/dbb/business/core/basic/Kibetsu.java
@@ -39,7 +39,7 @@ public class Kibetsu extends ModelBase<KibetsuIdentifier, DbT2003KibetsuEntity, 
     public Kibetsu(FlexibleYear 調定年度,
             FlexibleYear 賦課年度,
             TsuchishoNo 通知書番号,
-            Decimal 履歴番号,
+            int 履歴番号,
             RString 徴収方法,
             int 期) {
         requireNonNull(調定年度, UrSystemErrorMessages.値がnull.getReplacedMessage("調定年度"));
@@ -129,7 +129,7 @@ public class Kibetsu extends ModelBase<KibetsuIdentifier, DbT2003KibetsuEntity, 
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
