+++ DB/DBB/standard/business/src/main/java/jp/co/ndensan/reams/db/dbb/business/core/basic/KibetsuBuilder.java
@@ -81,7 +81,7 @@ public class KibetsuBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link KibetsuBuilder}
      */
-    public KibetsuBuilder set履歴番号(Decimal 履歴番号) {
+    public KibetsuBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
