+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KogakuGassanJikoFutanGakuMeisai.java
@@ -44,7 +44,7 @@ public class KogakuGassanJikoFutanGakuMeisai
             HokenshaNo 保険者番号,
             RString 支給申請書整理番号,
             RString 対象月,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         requireNonNull(被保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("被保険者番号"));
         requireNonNull(対象年度, UrSystemErrorMessages.値がnull.getReplacedMessage("対象年度"));
         requireNonNull(保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("保険者番号"));
@@ -150,7 +150,7 @@ public class KogakuGassanJikoFutanGakuMeisai
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
