+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KogakuGassanShinseishoKanyurekiBuilder.java
@@ -7,13 +7,12 @@ package jp.co.ndensan.reams.db.dbc.business.core.basic;
 
 import static java.util.Objects.requireNonNull;
 import jp.co.ndensan.reams.db.dbc.entity.db.basic.DbT3069KogakuGassanShinseishoKanyurekiEntity;
-import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HihokenshaNo;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HokenshaNo;
+import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 
 /**
  * {@link KogakuGassanShinseishoKanyureki}の編集を行うビルダークラスです。
@@ -107,7 +106,7 @@ public class KogakuGassanShinseishoKanyurekiBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link KogakuGassanShinseishoKanyurekiBuilder}
      */
-    public KogakuGassanShinseishoKanyurekiBuilder set履歴番号(Decimal 履歴番号) {
+    public KogakuGassanShinseishoKanyurekiBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
