+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShichosonTokubetuKyufuServiceBuilder.java
@@ -8,9 +8,9 @@ package jp.co.ndensan.reams.db.dbc.business.core.basic;
 import static java.util.Objects.requireNonNull;
 import jp.co.ndensan.reams.db.dbc.entity.db.basic.DbT3066ShichosonTokubetuKyufuServiceEntity;
 import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
+import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
 import jp.co.ndensan.reams.uz.uza.math.Decimal;
-import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 
 /**
  * {@link ShichosonTokubetuKyufuService}の編集を行うビルダークラスです。
@@ -68,7 +68,7 @@ public class ShichosonTokubetuKyufuServiceBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link ShichosonTokubetuKyufuServiceBuilder}
      */
-    public ShichosonTokubetuKyufuServiceBuilder set履歴番号(Decimal 履歴番号) {
+    public ShichosonTokubetuKyufuServiceBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
