+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/YoboKeikakuJikoSakuseiGokeiIdentifier.java
@@ -22,7 +22,7 @@ public class YoboKeikakuJikoSakuseiGokeiIdentifier implements Serializable {
 
     private final HihokenshaNo 被保険者番号;
     private final FlexibleYearMonth 対象年月;
-    private final Decimal 履歴番号;
+    private final int 履歴番号;
     private final RString 居宅サービス区分;
     private final JigyoshaNo サービス提供事業者番号;
     private final ServiceShuruiCode サービス種類コード;
@@ -39,7 +39,7 @@ public class YoboKeikakuJikoSakuseiGokeiIdentifier implements Serializable {
      */
     public YoboKeikakuJikoSakuseiGokeiIdentifier(HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号,
+            int 履歴番号,
             RString 居宅サービス区分,
             JigyoshaNo サービス提供事業者番号,
             ServiceShuruiCode サービス種類コード) {
