+++ DB/DBC/standard/business/src/test/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KogakuGassanShinseishoKanyurekiBuilderTest.java
@@ -4,14 +4,13 @@
  */
 package jp.co.ndensan.reams.db.dbc.business.core.basic;
 
-import jp.co.ndensan.reams.db.dbc.entity.db.basic.DbT3069KogakuGassanShinseishoKanyurekiEntity;
 import jp.co.ndensan.reams.db.dbc.entity.basic.helper.DbT3069KogakuGassanShinseishoKanyurekiEntityGenerator;
+import jp.co.ndensan.reams.db.dbc.entity.db.basic.DbT3069KogakuGassanShinseishoKanyurekiEntity;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HihokenshaNo;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HokenshaNo;
 import jp.co.ndensan.reams.db.dbz.testhelper.DbcTestBase;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import org.junit.Before;
@@ -34,7 +33,7 @@ public class KogakuGassanShinseishoKanyurekiBuilderTest extends DbcTestBase {
     private static HokenshaNo 主キー名3;
     private static RString 主キー名4;
     private static RString 主キー名5;
-    private static Decimal 主キー名6;
+    private static int 主キー名6;
 
     @BeforeClass
     public static void setUpClass() {
