+++ DB/DBC/standard/business/src/test/java/jp/co/ndensan/reams/db/dbc/business/core/basic/RiyoshaFutanWariaiMeisaiBuilderTest.java
@@ -4,12 +4,11 @@
  */
 package jp.co.ndensan.reams.db.dbc.business.core.basic;
 
-import jp.co.ndensan.reams.db.dbc.entity.db.basic.DbT3114RiyoshaFutanWariaiMeisaiEntity;
 import jp.co.ndensan.reams.db.dbc.entity.basic.helper.DbT3114RiyoshaFutanWariaiMeisaiEntityGenerator;
+import jp.co.ndensan.reams.db.dbc.entity.db.basic.DbT3114RiyoshaFutanWariaiMeisaiEntity;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HihokenshaNo;
 import jp.co.ndensan.reams.db.dbz.testhelper.DbcTestBase;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import org.junit.Before;
@@ -29,8 +28,8 @@ public class RiyoshaFutanWariaiMeisaiBuilderTest extends DbcTestBase {
 //TODO 主キーの数が足りない場合、追加してください。
     private static FlexibleYear 年度;
     private static HihokenshaNo 被保険者番号;
-    private static Decimal 履歴番号;
-    private static Decimal 枝番号;
+    private static int 履歴番号;
+    private static int 枝番号;
 
     @BeforeClass
     public static void setUpClass() {
