+++ DB/DBC/standard/business/src/test/java/jp/co/ndensan/reams/db/dbc/business/core/basic/RiyoshaFutanWariaiMeisaiIdentifierTest.java
@@ -9,7 +9,6 @@ import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HihokenshaN
 import static jp.co.ndensan.reams.db.dbx.testhelper.matcher.IsSerializable.serializable;
 import jp.co.ndensan.reams.db.dbz.testhelper.DbcTestBase;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import org.junit.BeforeClass;
@@ -27,8 +26,8 @@ public class RiyoshaFutanWariaiMeisaiIdentifierTest extends DbcTestBase {
 //TODO 主キーの数が足りない場合、追加してください。
     private static FlexibleYear 年度;
     private static HihokenshaNo 被保険者番号;
-    private static Decimal 履歴番号;
-    private static Decimal 枝番号;
+    private static int 履歴番号;
+    private static int 枝番号;
 
     @BeforeClass
     public static void setUpClass() {
